+++ Plugins/WorldAssembly/Source/NexusWorldAssembly/Public/Cell/NTissueTagGroups.h
@@ -42,10 +42,24 @@ struct FNTissueTagGroups
 		MustHaveTags.AppendTags(OtherTagGroup.MustHaveTags);
 	}
 	
-protected:
-	UPROPERTY(EditAnywhere, DisplayName="Unique", meta = (Categories="NEXUS.WorldAssembly"))
-	FGameplayTagContainer UniqueTags = FGameplayTagContainer(NWorldAssembly_BuiltIn_Unique);
+	void EnsureBaseTags()
+	{
+		if (!UniqueTags.HasTag(NWorldAssembly_BuiltIn_Unique))
+		{
+			UniqueTags.AppendTags(FGameplayTagContainer(NWorldAssembly_BuiltIn_Unique));
+		}
+		if (!MustHaveTags.HasTag(NWorldAssembly_BuiltIn_MustHave))
+		{
+			MustHaveTags.AppendTags(FGameplayTagContainer(NWorldAssembly_BuiltIn_MustHave));
+		}
+	}
 	
-	UPROPERTY(EditAnywhere, DisplayName="Must Have", meta = (Categories="NEXUS.WorldAssembly"))
-	FGameplayTagContainer MustHaveTags = FGameplayTagContainer(NWorldAssembly_BuiltIn_MustHave);
+protected:
+	UPROPERTY(EditAnywhere, DisplayName="Unique", meta = (Categories="NEXUS.WorldAssembly", 
+		ToolTip="When a tag group is flagged as unique it means that only one item can be placed in the graph from that group. NEXUS.WorldAssembly.BuiltIn.Unique is auto-added to this group. NOTE: If combined with Must Have, the MinimumCount will be ignored for all in the group, just ensuring that one is present."))
+	FGameplayTagContainer UniqueTags;
+	
+	UPROPERTY(EditAnywhere, DisplayName="Must Have", meta = (Categories="NEXUS.WorldAssembly", 
+		ToolTip="When a tag group is marked Must Have, it means that when generating we require something from that group to be placed in the graph. NEXUS.WorldAssembly.BuiltIn.MustHave is auto-added to this group. NOTE: If combined with Unique, the MinimumCount will be ignored for all in the group, just ensuring that one is present."))
+	FGameplayTagContainer MustHaveTags;
 };